requirements.txt
amqp==2.3.2
attrs==18.1.0
Automat==0.7.0
Babel==2.6.0
billiard==3.5.0.3
celery==4.2.0
constantly==15.1.0
flower==0.9.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
kombu==4.2.1
Pillow==5.1.0
psycopg2-binary==2.7.5
pymongo==3.7.0
pypng==0.0.18
pytz==2018.5
redis==2.10.6
six==1.11.0
tornado==5.0.2
Twisted==18.4.0
txmongo==18.1.0
txpostgres==1.6.0
txredisapi==1.4.4
vine==1.1.4
zope.interface==4.5.0
